cvrp_project/cvrp_app/urls.py
from django import urls
from django.db import router
from django.urls import path, include
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

from rest_framework import routers
from .views import prueba, tipoDocumentoViewSet

router= routers.DefaultRouter()
router.register(r'docs',tipoDocumentoViewSet)

urlpatterns = [
    path('', include(router.urls)),
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh', TokenRefreshView.as_view(), name='token_refresh'),
]
