cvrp_project/requirements.txt
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
isort==5.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
mysql==0.0.2
mysqlclient @ file:///C:/xampp/htdocs/python/proyecto_solo/mysqlclient-1.4.6-cp38-cp38-win32.whl
PyJWT==2.0.1
pylint==2.6.2
PyMySQL==1.0.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
